+++ src/generator/challenges.rs
@@ -1,3 +1,5 @@
+use super::{Generator, G2331, G2332};
+use log::debug;
 use std::fmt::Display;
 
 #[derive(Debug, Clone)]
@@ -23,14 +25,18 @@ impl Challenges {
         let c2331 = Challenge {
             name: String::from("Find the odd one out"),
             command: String::from("2331"),
-            doc_path: String::from("/home/philip/code_challenges/judge_23_3_1/README.md"),
+            doc_path: String::from(
+                "/home/philip/code_challenges/judge_23_3_1/src/generator/2331.md",
+            ),
             table: String::from("23_3_1"),
         };
 
         let c2332 = Challenge {
             name: String::from("Find the odd one out two"),
             command: String::from("2332"),
-            doc_path: String::from("/home/philip/code_challenges/judge_23_3_1/README.md"),
+            doc_path: String::from(
+                "/home/philip/code_challenges/judge_23_3_1/src/generator/2332.md",
+            ),
             table: String::from("23_3_2"),
         };
 
@@ -41,6 +47,15 @@ impl Challenges {
 
     pub fn get_challenge(&self, challenge: &str) -> Option<&Challenge> {
         self.challenges.iter().find(|c| c.command == challenge)
+    }
+
+    pub fn make_generator(&self, challenge: &str, count: usize) -> Option<Box<dyn Generator>> {
+        debug!("make_generator: {}", challenge);
+        match challenge {
+            "2331" => Some(Box::new(G2331::new(count))),
+            "2332" => Some(Box::new(G2332::new(count))),
+            _ => None,
+        }
     }
 }