+++ src/generator/test_result.rs
@@ -1,0 +1,8 @@
+use scoreboard_db::Score;
+
+#[derive(Clone)]
+pub enum TestResult {
+    Success(Score),
+    Fail(String),
+    Stolen(Score, Score),
+}